// reportus-backend/src/payload-types.ts
export interface User {
  id: string;
  role: 'admin' | 'user_general' | 'user_soc' | 'manager_soc' | 'manager_general';
  username: string;
  first_name: string;
  last_name: string;
  phone?: string | null;
  avatar?: (string | null) | Media;
  department:
    | 'Geschäftsführung'
    | 'Netzwerk'
    | 'Server'
    | 'Entwicklung'
    | 'SOC'
    | 'Vertrieb'
    | 'Buchhaltung'
    | 'Controling'
    | 'Generic';
  _status?: ('active' | 'archived') | null;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  _verified?: boolean | null;
  _verificationToken?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  password?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: string;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}